+++ tests/Tests/Unit/GapicClientTraitTest.php
@@ -41,6 +41,7 @@ use Google\ApiCore\CredentialsWrapper;
 use Google\ApiCore\GapicClientTrait;
 use Google\ApiCore\LongRunning\OperationsClient;
 use Google\ApiCore\OperationResponse;
+use Google\ApiCore\Options\TransportOptions;
 use Google\ApiCore\RequestParamsHeaderDescriptor;
 use Google\ApiCore\RetrySettings;
 use Google\ApiCore\ServerStream;
@@ -866,7 +867,7 @@ class GapicClientTraitTest extends TestCase
         }
         $expectedProperties = [
             'serviceName' => 'test.interface.v1.api',
-            'agentHeader' => AgentHeader::buildAgentHeader([]),
+            'agentHeader' => AgentHeader::buildAgentHeader([]) + ['User-Agent' => ['gcloud-php-legacy/']],
             'retrySettings' => $expectedRetrySettings,
         ];
         return [
@@ -1118,7 +1119,7 @@ class GapicClientTraitTest extends TestCase
         $this->assertTrue($updatedOptions['disableRetries']);
     }
 
-    private function buildClientToTestModifyCallMethods()
+    private function buildClientToTestModifyCallMethods($clientClass = null)
     {
         $header = AgentHeader::buildAgentHeader([]);
         $retrySettings = $this->getMockBuilder(RetrySettings::class)
@@ -1145,7 +1146,8 @@ class GapicClientTraitTest extends TestCase
         $credentialsWrapper = CredentialsWrapper::build([
             'keyFile' => __DIR__ . '/testdata/json-key-file.json'
         ]);
-        $client = new GapicClientTraitStubExtension();
+        $clientClass = $clientClass ?: GapicClientTraitStubExtension::class;
+        $client = new $clientClass();
         $client->set('transport', $transport);
         $client->set('credentialsWrapper', $credentialsWrapper);
         $client->set('agentHeader', $header);
@@ -1350,7 +1352,8 @@ class GapicClientTraitTest extends TestCase
                 $this->isInstanceOf(Call::class),
                 $this->equalTo([
                     'headers' => AgentHeader::buildAgentHeader([]) + [
-                        'X-Goog-User-Project' => [$quotaProject]
+                        'X-Goog-User-Project' => [$quotaProject],
+                        'User-Agent' => ['gcloud-php-legacy/']
                     ],
                     'credentialsWrapper' => $credentialsWrapper
                 ])
@@ -1566,13 +1569,13 @@ class GapicClientTraitTest extends TestCase
     public function testSupportedTransportOverrideWithDefaultTransport()
     {
         $client = new GapicClientTraitRestOnly();
-        $this->assertInstanceOf(RestTransport::class, $client->getTransport());
+        $this->assertInstanceOf(RestTransport::class, $client->call('getTransport'));
     }
 
     public function testSupportedTransportOverrideWithExplicitTransport()
     {
         $client = new GapicClientTraitRestOnly(['transport' => 'rest']);
-        $this->assertInstanceOf(RestTransport::class, $client->getTransport());
+        $this->assertInstanceOf(RestTransport::class, $client->call('getTransport'));
     }
 
     /** @dataProvider provideDetermineMtlsEndpoint */
@@ -1712,6 +1715,124 @@ class GapicClientTraitTest extends TestCase
         $this->assertTrue(is_callable($options['clientCertSource']));
         $this->assertEquals(['foo', 'foo'], $options['clientCertSource']());
     }
+
+    public function testInvalidClientOptionsTypeThrowsExceptionForV2SurfaceOnly()
+    {
+        // v1 client
+        new GapicClientTraitStub(['apiEndpoint' => ['foo']]);
+        $this->assertTrue(true, 'Test made it to here without throwing an exception');
+
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage(
+            PHP_MAJOR_VERSION < 8
+                ? 'Argument 1 passed to Google\ApiCore\Options\ClientOptions::setApiEndpoint() '
+                    . 'must be of the type string or null, array given'
+                : 'Google\ApiCore\Options\ClientOptions::setApiEndpoint(): Argument #1 '
+                    . '($apiEndpoint) must be of type ?string, array given'
+        );
+
+        // v2 client
+        new GapicV2SurfaceClient(['apiEndpoint' => ['foo']]);
+    }
+
+    public function testCallOptionsForV2Surface()
+    {
+        list($client, $transport) = $this->buildClientToTestModifyCallMethods(
+            GapicV2SurfaceClient::class
+        );
+
+        $transport->expects($this->once())
+            ->method('startUnaryCall')
+            ->with(
+                $this->isInstanceOf(Call::class),
+                $this->equalTo([
+                    'headers' => AgentHeader::buildAgentHeader([]) + ['Foo' => 'Bar'],
+                    'credentialsWrapper' => CredentialsWrapper::build([
+                        'keyFile' => __DIR__ . '/testdata/json-key-file.json'
+                    ]),
+                    'timeoutMillis' => null, // adds null timeoutMillis
+                ])
+            )
+            ->willReturn(new FulfilledPromise(new Operation()));
+
+        $callOptions = [
+            'headers' => ['Foo' => 'Bar'],
+            'invalidOption' => 'wont-be-passed'
+        ];
+        $client->call('startCall', [
+            'simpleMethod',
+            'decodeType',
+            $callOptions,
+            new MockRequest(),
+        ])->wait();
+    }
+
+    public function testInvalidCallOptionsTypeForV1SurfaceDoesNotThrowException()
+    {
+        list($client, $transport) = $this->buildClientToTestModifyCallMethods();
+
+        $transport->expects($this->once())
+            ->method('startUnaryCall')
+            ->with(
+                $this->isInstanceOf(Call::class),
+                $this->equalTo([
+                    'transportOptions' => ['custom' => ['addModifyUnaryCallableOption' => true]],
+                    'headers' => AgentHeader::buildAgentHeader([]),
+                    'credentialsWrapper' => CredentialsWrapper::build([
+                        'keyFile' => __DIR__ . '/testdata/json-key-file.json'
+                    ]),
+                    'timeoutMillis' => 'blue', // invalid type, this is ignored
+                ])
+            )
+            ->willReturn(new FulfilledPromise(new Operation()));
+
+        $client->call('startCall', [
+            'simpleMethod',
+            'decodeType',
+            ['timeoutMillis' => 'blue'],
+            new MockRequest(),
+        ])->wait();
+    }
+
+    public function testInvalidCallOptionsTypeForV2SurfaceThrowsException()
+    {
+        $this->expectException(\TypeError::class);
+        $this->expectExceptionMessage(
+            PHP_MAJOR_VERSION < 8
+                ? 'Argument 1 passed to Google\ApiCore\Options\CallOptions::setTimeoutMillis() '
+                    . 'must be of the type int or null, string given'
+                : 'Google\ApiCore\Options\CallOptions::setTimeoutMillis(): Argument #1 '
+                    . '($timeoutMillis) must be of type ?int, string given'
+        );
+
+        list($client, $_) = $this->buildClientToTestModifyCallMethods(GapicV2SurfaceClient::class);
+
+        $client->call('startCall', [
+            'simpleMethod',
+            'decodeType',
+            ['timeoutMillis' => 'blue'], // invalid type, will throw exception
+            new MockRequest(),
+        ])->wait();
+    }
+
+    public function testSurfaceAgentHeaders()
+    {
+        // V1 does not contain new headers
+        $client = new GapicClientTraitRestOnly([
+            'gapicVersion' => '0.0.2',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.2 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-legacy/0.0.2', $agentHeader['User-Agent'][0]);
+
+        // V2 contains new headers
+        $client = new GapicV2SurfaceClient([
+            'gapicVersion' => '0.0.1',
+        ]);
+        $agentHeader = $client->getAgentHeader();
+        $this->assertStringContainsString(' gapic/0.0.1 ', $agentHeader['x-goog-api-client'][0]);
+        $this->assertEquals('gcloud-php-new/0.0.1', $agentHeader['User-Agent'][0]);
+    }
 }
 
 class GapicClientTraitStub
@@ -1754,26 +1875,6 @@ class GapicClientTraitStub
                 ]
             ],
         ];
-    }
-
-    public function set($name, $val, $static = false)
-    {
-        if (!property_exists($this, $name)) {
-            throw new \InvalidArgumentException("Property not found: $name");
-        }
-        if ($static) {
-            $this::$$name = $val;
-        } else {
-            $this->$name = $val;
-        }
-    }
-
-    public function get($name)
-    {
-        if (!property_exists($this, $name)) {
-            throw new \InvalidArgumentException("Property not found: $name");
-        }
-        return $this->$name;
     }
 }
 
@@ -1851,6 +1952,7 @@ class GapicClientTraitRestOnly
 {
     use GapicClientTrait {
         buildClientOptions as public;
+        getTransport as public;
     }
 
     public function __construct($options = [])
@@ -1874,11 +1976,6 @@ class GapicClientTraitRestOnly
         ];
     }
 
-    public function getTransport()
-    {
-        return $this->transport;
-    }
-
     private static function supportedTransports()
     {
         return ['rest', 'fake-transport'];
@@ -1887,6 +1984,11 @@ class GapicClientTraitRestOnly
     private static function defaultTransport()
     {
         return 'rest';
+    }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
     }
 }
 
@@ -1905,4 +2007,40 @@ class CustomOperationsClient
     public function getOperation($name, $arg1, $arg2)
     {
     }
+}
+
+abstract class GapicV2SurfaceBaseClient
+{
+    use GapicClientTrait;
+    use GapicClientStubTrait;
+
+    public function __construct(array $options = [])
+    {
+        $clientOptions = $this->buildClientOptions($options);
+        $this->setClientOptions($clientOptions);
+    }
+
+    public static function getClientDefaults()
+    {
+        return [
+            'apiEndpoint' => 'test.address.com:443',
+            'serviceName' => 'test.interface.v1.api',
+            'clientConfig' => __DIR__ . '/testdata/test_service_client_config.json',
+            'descriptorsConfigPath' => __DIR__.'/testdata/test_service_descriptor_config.php',
+            'transportConfig' => [
+                'rest' => [
+                    'restClientConfigPath' => __DIR__.'/testdata/test_service_rest_client_config.php',
+                ]
+            ],
+        ];
+    }
+
+    public function getAgentHeader()
+    {
+        return $this->agentHeader;
+    }
+}
+
+class GapicV2SurfaceClient extends GapicV2SurfaceBaseClient
+{
 }